+++ ext/scripts/parse_ansi.js
@@ -1274,7 +1274,6 @@ class Markup {
    * @param [html=``] string
    */
   _tagBlockCharacters(html = ``) {
-    if (BrowserOS() === Os.macOS) return html
     const config = localStorage.getItem(`textSmearBlockCharacters`) || `false`
     if (config !== `true`) return html
     return html.replace(RegExp(/([◘░▒▓█▄▐▌▀■]+)/, `g`), `<b>$1</b>`)
@@ -2302,5 +2301,5 @@ class Build extends Scan {
 }
 
 /*eslint no-useless-escape: "warn"*/
-/*global BrowserOS CheckArguments CheckError CheckRange Os */
+/*global CheckArguments CheckError CheckRange */
 /*exported Controls*/