+++ docs/changes.md
@@ -33,6 +33,7 @@ hide:
 - Mentions of the file scheme `file:///` will under Windows display as `file:///C:/`
 - Updated the Welcome, new install text to be a Getting started with RetroTxt brief.
 - Monitor downloads works better with 16colo.rs by ignoring their incorrect `Content-Type` headers.
+- Smear block characters are applied to BBS texts.
 
 * Using pnpm as the dependencies manager.
 * Removed the Firefox specific build tools.